Renamed module to palantirjavaformat
