+++ contrib/package.mill
@@ -241,7 +241,7 @@ object `package` extends RootModule {
     def testModuleDeps = super.testModuleDeps ++ Seq(build.scalalib)
   }
 
-  object palantir extends ContribModule {
+  object palantirjavaformat extends ContribModule {
     def compileModuleDeps = Seq(build.scalalib)
     def testModuleDeps = super.testModuleDeps ++ Seq(build.scalalib)
   }
