+++ contrib/palantirjavaformat/src/mill/contrib/palantirjavaformat/PalantirJavaFormatOptions.scala
@@ -1,5 +1,5 @@
 package mill
-package contrib.palantir
+package contrib.palantirjavaformat
 
 /**
  * @param styleOverride override default style with `palantir | asop`
@@ -7,15 +7,15 @@ package contrib.palantir
  * @param skipUnusedImports do not remove unused imports
  * @param skipReflowingLongStrings do not wrap long strings
  */
-case class PalantirOptions(
+case class PalantirJavaFormatOptions(
     styleOverride: Option[String] = Some("palantir"),
     skipSortImports: Boolean = false,
     skipUnusedImports: Boolean = false,
     skipReflowingLongStrings: Boolean = false
 )
-object PalantirOptions {
+object PalantirJavaFormatOptions {
 
   import upickle.default._
 
-  implicit val RW: ReadWriter[PalantirOptions] = macroRW[PalantirOptions]
+  implicit val RW: ReadWriter[PalantirJavaFormatOptions] = macroRW[PalantirJavaFormatOptions]
 }
