+++ contrib/palantirjavaformat/test/expected/palantir/Main.java
