+++ contrib/palantirjavaformat/test/resources/src/Main.java
